+++ apps/Admin.js
@@ -239,7 +239,7 @@ export default class UCAdmin extends UCPlugin {
       if (setData) {
         log.debug(setData)
         const [g, s] = setData.path.split('.')
-        if (isGlobal || (s && !isGlobal && s in UCPr.groupCFG(groupId)[type][g])) {
+        if (isGlobal || type === 'GAconfig' || (s && !isGlobal && s in UCPr.groupCFG(groupId)[type][g])) {
           showGroup = group
           // 获取新设置值
           let operation