+++ apps/addJS.js
@@ -37,7 +37,7 @@ export default class UCAddJS extends UCPlugin {
   }
 
   async _fileContext() {
-    if (this.isCancel()) return false
+    if (this.isCancel()) return
     if (!this.isGroup && !this.e.friend) return this.finishReply('请先添加好友')
     if (!this.e.file) return this.reply('请发送js文件')
     else {
@@ -51,7 +51,7 @@ export default class UCAddJS extends UCPlugin {
         this.e.isUCJS = isUCJS
         this.e.filename = filename
         this.e.dirPath = dirPath
-        this.finish(this.setFnc)
+        this.finishUCcontext(this.setFnc)
         this.setUCcontext(this.setFnc2)
         return this.reply(`你已经安装过${dirPath}${filename}插件了，是否覆盖原插件？[是|否]`)
       }
@@ -62,7 +62,7 @@ export default class UCAddJS extends UCPlugin {
       } else {
         this.errorReply()
       }
-      this.finish(this.setFnc)
+      this.finishUCcontext(this.setFnc)
     }
   }
 
@@ -80,7 +80,7 @@ export default class UCAddJS extends UCPlugin {
       } else {
         this.errorReply()
       }
-      return this.finish(this.setFnc2)
+      return this.finishUCcontext(this.setFnc2)
     }
   }