+++ models/UCEvent.js
@@ -1,7 +1,7 @@
-import { UCPr, file, Path, Check, common } from '../components/index.js'
+import { UCPr, file, Path, common } from '../components/index.js'
 import Runtime from '../../../lib/plugins/runtime.js'
 import loader from '../../../lib/plugins/loader.js'
-import { UCPlugin } from './index.js'
+import UCPlugin from './UCPlugin.js'
 import _ from 'lodash'
 
 const hook = []
@@ -28,75 +28,12 @@ export default class UCEvent extends UCPlugin {
      * - `request.group`: invite add
      */
     this.sub_type = ''
-    if (!e) return
-    /** bot是否为管理员 */
-    this.botIsAdmin = this.e.group?.is_admin
-    /** bot是否为群主 */
-    this.botIsOwner = this.e.group?.is_owner
-    /** bot是否管理员或群主 */
-    this.botIsAdminOrOwner = this.botIsAdmin || this.botIsOwner
-    /** 群名 */
-    this.groupName = this.e.group?.name
-  }
-
-  /** 是否启用群管及该功能 */
-  get isOpen() {
-    return this.GAconfig?.isOpen && _.get(this.Cfg, 'isOpen', true)
-  }
-
-  /** 检查是否全局主人 */
-  isGM(userId) {
-    return Check.Master(userId)
-  }
-
-  /** 检查是否主人 */
-  isM(userId, groupId) {
-    if (this.isGM(userId)) return true
-    if (groupId) return Check.str(UCPr.groupCFG(groupId).permission?.Master, userId)
-    return Check.str(this.groupCFG.permission?.Master, userId)
-  }
-
-  /** 检查是否全局管理 */
-  isGA(userId) {
-    return Check.Admin(userId)
-  }
-
-  /** 检查是否管理 */
-  isA(userId, groupId) {
-    if (this.isGA(userId)) return true
-    if (groupId) return Check.str(UCPr.groupCFG(groupId).permission?.Admin, userId)
-    return Check.str(this.groupCFG.permission?.Admin, userId)
-  }
-
-  /** 检查开关、use权限?、Bot群权限 */
-  defaultVerify(isVerifyPermission = true) {
-    if (!this.isOpen) return false
-    if (isVerifyPermission && !this.verifyPermission(this.Cfg.use)) return false
-    if (!this.botIsAdminOrOwner) return this.noPowReply()
-    return true
   }
 
   getMemInfo(memId) {
     const memClient = this.e.group.pickMember(memId)
     if (!memClient?.info) return null
     return memClient.info
-  }
-
-  /** Bot和待操作群员群权限对比 */
-  checkBotPower(memId = this.userId) {
-    if (this.botIsOwner) return true
-    if (memId == this.qq) return true
-    if (!this.botIsAdmin) return false
-    if (!memId) return false
-    const memInfo = this.getMemInfo(memId)
-    if (!memInfo) return false
-    if (memInfo.role === 'admin' || memInfo.role === 'owner') return false
-    return true
-  }
-
-  /** 用户和待操作群员插件权限对比 */
-  checkUserPower(memId) {
-    return this.level >= Check.level(memId, this.groupId)
   }
 
   getMemName(memId) {
@@ -107,7 +44,7 @@ export default class UCEvent extends UCPlugin {
   }
 
   /** 创建上下文，返回false再处理 */
-  setUCcontext(fnc = this.setFnc || '__chooseContext', time = UCPr.GAconfig.overTime) {
+  setUCcontext(fnc = this.setFnc, time = UCPr.GAconfig.overTime) {
     const key = this.conKey()
     if (_.some(hook, { key })) return
     const info = {
@@ -170,13 +107,13 @@ class UCSwitchBotEvent extends UCEvent {
 }
 
 async function UCdealMsg(type, e) {
+  const msg = _.filter(e.message, { type: 'text' }).map(v => v.text).join(' ').trim()
   const groupId = e.group_id || e.group?.gid
   // 是群内并且UC下线了Bot
-  if (groupId && _.isEqual(_.get(UCPr.defaultCfg.getConfig('group'), `${groupId}.enable`), ['UC-switchBot'])) {
+  if (msg && groupId && _.isEqual(_.get(UCPr.defaultCfg.getConfig('group'), `${groupId}.enable`), ['UC-switchBot'])) {
     const a = new UCSwitchBotEvent(e)
     return a.deal(e, type)
   }
-  const msg = _.filter(e.message, { type: 'text' }).map(v => v.text).join(' ').trim()
   const events = UCPr.event[type].filter(({ sub_type }) => sub_type === 'all' || sub_type === e.sub_type)
   /** 处理hook */
   for (const event of events) {
@@ -220,8 +157,10 @@ async function UCdealMsg(type, e) {
         // log.white(msg)
         const start = Date.now()
         const logInfo = `[${event.name}][${rule.fnc}]${_.truncate(msg, { length: 50 })}`
-        log.white(logInfo)
         const app = new event.class(e)
+        // 开启了UC前缀
+        if (msg && app.isUCGA && app.GAconfig.isPrefix && !app.checkPrefix()) return false
+        log.white(logInfo)
         const result = await app[rule.fnc](e)?.catch?.(err => {
           const errInfo = log.error(`执行${logInfo}错误`, err)
           e.reply?.(errInfo)