+++ models/index.js
@@ -1,8 +1,9 @@
 import UCPlugin from './UCPlugin.js'
 import UCEvent from './UCEvent.js'
+import UCGAPlugin from './UCGAPlugin.js'
 import Permission from './Permission.js'
 import Help from './Help.js'
 import Cfg from './Cfg.js'
 import Sqtj from './Sqtj.js'
 
-export { UCPlugin, UCEvent, Permission, Help, Cfg, Sqtj }
+export { UCPlugin, UCEvent, UCGAPlugin, Permission, Help, Cfg, Sqtj }